+++ src/widget/image_button_widget.rs
@@ -188,6 +188,10 @@ impl Widget for ImageButtonWidget {
         None
     }
 
+    fn handles_events(&mut self) -> bool {
+        true
+    }
+
     fn set_widget_id(&mut self, widget_id: i32) {
         self.widget_id = widget_id;
     }