+++ src/widget/push_button_widget.rs
@@ -152,6 +152,10 @@ impl Widget for PushButtonWidget {
         None
     }
 
+    fn handles_events(&mut self) -> bool {
+        true
+    }
+
     fn set_widget_id(&mut self, widget_id: i32) {
         self.widget_id = widget_id;
     }